+++ changelog.md
@@ -1,5 +1,6 @@
 
 ## TODO
+	MIDI notation from LilyDocument, MIDI matching by LilyInterpreter
 	implicit notes matching with sheet document
 		try articulate.ly in MIDI output
 	staffNotation